+++ app/conf/__init__.py
@@ -4,7 +4,7 @@ import logging
 env = environ.get('APP_ENV') or 'development'
 logging.info("APP_ENV = " + env)
 
-from conf import s3, geolocation, spocs, adzerk, domain_affinities
+from app.conf import s3, geolocation, spocs, adzerk, domain_affinities
 
 s3 = getattr(s3, env)
 geolocation = getattr(geolocation, env)
