+++ app/conf/domain_affinities.py
