+++ app/exceptions/invalid_content_type.py
@@ -1,4 +1,5 @@
 from app.exceptions.base_exception import BaseException
 
+
 class InvalidContentType(BaseException):
     status_code = 400
