+++ app/exceptions/invalid_param.py
@@ -1,4 +1,5 @@
 from app.exceptions.base_exception import BaseException
 
+
 class InvalidParam(BaseException):
     status_code = 400
