+++ conftest.py
@@ -1,0 +1,2 @@
+# This file must exist for pytest to work properly.
+# https://stackoverflow.com/a/50610630/732649
