+++ tests/unit/test_adzerk_transform.py
@@ -1,41 +1,40 @@
 from unittest import TestCase
 from unittest.mock import patch
 
-from adzerk.transform import \
+from app.adzerk.transform import \
     to_spoc, tracking_url_to_shim, is_collection, to_collection, get_personalization_models
 from tests.fixtures.mock_spoc import *
 from tests.fixtures.mock_decision import *
-import conf
 
 
 class TestAdZerkTransform(TestCase):
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc(self):
         self.assertEqual(mock_spoc_2, to_spoc(mock_decision_2))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc_cta(self):
         self.assertEqual(mock_spoc_3_cta, to_spoc(mock_decision_3_cta))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc_topics(self):
         self.assertEqual(mock_spoc_5_topics, to_spoc(mock_decision_5_topics))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc_no_sponsor(self):
         self.assertEqual(mock_spoc_6_no_sponsor, to_spoc(mock_decision_6_no_sponsor))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc_is_video(self):
         self.assertEqual(mock_spoc_7_is_video, to_spoc(mock_decision_7_is_video))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_to_spoc_sponsored_by_override(self):
         self.assertEqual(mock_spoc_8_blank_sponsored_by_override, to_spoc(mock_decision_8_blank_sponsored_by_override))
         self.assertEqual(mock_spoc_9_sponsored_by_override, to_spoc(mock_decision_9_sponsored_by_override))
 
-    @patch.dict('conf.domain_affinities', {"publishers": {'example.com': 1}})
+    @patch.dict('app.conf.domain_affinities', {"publishers": {'example.com': 1}})
     def test_missing_priority_id(self):
         self.assertEqual(mock_spoc_10_missing_priority, to_spoc(mock_decision_10_missing_priority))
