+++ tests/unit/test_app.py
@@ -39,7 +39,7 @@ class TestApp(unittest.TestCase):
     Tests: Pulse
     """
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
     def test_app_pulse(self, mock_geo):
         resp = self.create_client_no_geo_locs().get('/pulse')
         self.assertEqual(resp.json(), {"pulse" : "ok"})
@@ -48,14 +48,14 @@ class TestApp(unittest.TestCase):
     Tests: spocs
     """
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_valid(self, mock_geo, mock_adzerk):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body())
         self.assertEqual(resp.status_code, 200)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_valid_with_country_region(self, mock_geo, mock_adzerk):
         country_region = {
             'country': 'CA',
@@ -64,8 +64,8 @@ class TestApp(unittest.TestCase):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(update=country_region))
         self.assertEqual(resp.status_code, 200)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_collection_placement_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_collection_placement_map)
     def test_app_spocs_collection_v1(self, mock_geo, mock_adzerk):
         """
         API version 1 returns the collection as an array for backwards compatibility.
@@ -75,8 +75,8 @@ class TestApp(unittest.TestCase):
         self.assertEqual(resp.status_code, 200)
         self.assertEqual(resp.json()['sponsored-collection'][0]['collection_title'], 'Best of the Web')
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_collection_placement_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_collection_placement_map)
     def test_app_spocs_collection_v2(self, mock_geo, mock_adzerk):
         """
         API version 2 returns the collection as an object, with collection-level fields pulled up.
@@ -93,42 +93,42 @@ class TestApp(unittest.TestCase):
         self.assertEqual(collection['items'][0]['title'], 'title 900')
         self.assertTrue('collection_title' not in collection['items'][0])
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_invalid_no_version(self, mock_geo, mock_adzerk):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(without='version'))
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_invalid_no_pocket_id(self, mock_geo, mock_adzerk):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(without='pocket_id'))
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_invalid_no_consumer_key(self, mock_geo, mock_adzerk):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(without='consumer_key'))
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_invalid_pocket_id(self, mock_geo, mock_adzerk):
         data = self.get_request_body()
         data['pocket_id'] = 'invalid'
         resp = self.create_client_no_geo_locs().post('/spocs', json=data)
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_unrecognized_field(self, mock_geo, mock_adzerk):
         data = self.get_request_body()
         data['invalid'] = 'something'
         resp = self.create_client_no_geo_locs().post('/spocs', json=data)
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_production_invalid_content_type(self, mock_geo, mock_adzerk):
         resp = self.create_client_no_geo_locs().post(
             '/spocs',
@@ -137,8 +137,8 @@ class TestApp(unittest.TestCase):
         )
         self.assertEqual(resp.status_code, 400)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_response_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_response_map)
     def test_app_spocs_staging_production_valid(self, mock_geo, mock_adzerk):
         """
         This test would be more useful if we checked that we got different responses based on the site.
@@ -150,27 +150,27 @@ class TestApp(unittest.TestCase):
         resp = self.create_client_no_geo_locs().post('/spocs?site=12345', json=self.get_request_body())
         self.assertEqual(resp.status_code, 200)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
     def test_app_spocs_production_valid_placements(self, mock_geo):
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(placements=mock_placements))
         self.assertEqual(200, resp.status_code)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
     def test_app_spocs_production_invalid_placement_name(self, mock_geo):
         bad_placements = deepcopy(mock_placements)
         bad_placements[0].pop('name')
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(placements=bad_placements))
         self.assertEqual(400, resp.status_code)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
     def test_app_spocs_production_unknown_placement_field(self, mock_geo):
         bad_placements = deepcopy(mock_placements)
         bad_placements[0]['adTypess'] = ['test']
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(placements=bad_placements))
         self.assertEqual(400, resp.status_code)
 
-    @patch('provider.geo_provider.GeolocationProvider.__init__', return_value=None)
-    @patch('adzerk.api.Api.get_decisions', return_value=mock_placement_map)
+    @patch('app.provider.geo_provider.GeolocationProvider.__init__', return_value=None)
+    @patch('app.adzerk.api.Api.get_decisions', return_value=mock_placement_map)
     def test_app_spocs_production_valid_placements_call(self, mock_geo, mock_dec):
         bad_placements = deepcopy(mock_placements)
         resp = self.create_client_no_geo_locs().post('/spocs', json=self.get_request_body(placements=bad_placements))
