+++ tests/unit/test_geolocation_provider.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import patch
 
-from provider.geo_provider import GeolocationProvider
+from app.provider.geo_provider import GeolocationProvider
 from tests.fixtures.mock_factory import get_mocked_geolocation_factory
 
 
@@ -10,7 +10,7 @@ class TestGeolocationProvider(TestCase):
     __FACTORY = get_mocked_geolocation_factory()
 
     @patch('boto3.session.Session.client', return_value=None)
-    @patch('geolocation.factory.Factory.get_instance', return_value=__FACTORY)
+    @patch('app.geolocation.factory.Factory.get_instance', return_value=__FACTORY)
     def test_geolocation_no_setattr(self, mock_s3, mock_geofactory):
         glp = GeolocationProvider()
         try:
@@ -20,7 +20,7 @@ class TestGeolocationProvider(TestCase):
             self.assertEqual(e.__str__(), 'Already instantiated')
 
     @patch('boto3.session.Session.client', return_value=None)
-    @patch('geolocation.factory.Factory.get_instance', return_value=__FACTORY)
+    @patch('app.geolocation.factory.Factory.get_instance', return_value=__FACTORY)
     def test_geolocation_valid_ip(self, mock_s3, mock_geofactory):
         glp = GeolocationProvider()
         city = glp.get_city('216.160.83.56')
@@ -30,7 +30,7 @@ class TestGeolocationProvider(TestCase):
         self.assertEqual('WA', glp.get_region(city))
 
     @patch('boto3.session.Session.client', return_value=None)
-    @patch('geolocation.factory.Factory.get_instance', return_value=__FACTORY)
+    @patch('app.geolocation.factory.Factory.get_instance', return_value=__FACTORY)
     def test_geolocation_invalid_ip(self, mock_s3, mock_geofactory):
         glp = GeolocationProvider()
         try:
@@ -39,7 +39,7 @@ class TestGeolocationProvider(TestCase):
             self.assertEqual('The address 127.0.0.1 is not in the database.', e.__str__())
 
     @patch('boto3.session.Session.client', return_value=None)
-    @patch('geolocation.factory.Factory.get_instance', return_value=__FACTORY)
+    @patch('app.geolocation.factory.Factory.get_instance', return_value=__FACTORY)
     def test_geolocation_no_city(self, mock_s3, mock_geofactory):
         glp = GeolocationProvider()
         self.assertIsNone(glp.get_country(None))
